+++ great_international/serializers.py
@@ -514,6 +514,8 @@ class InternationalSectorPageSerializer(
     section_three_subsection_two_body = core_fields.MarkdownToHTMLField()
 
     project_opportunities_title = serializers.CharField(max_length=255)
+    related_opportunities_cta_text = serializers.CharField(max_length=255)
+    related_opportunities_cta_link = serializers.CharField(max_length=255)
 
     related_opportunities = serializers.SerializerMethodField()